misc/HtmlTestSuite/results/test-1023.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Canvas test: 2d.pattern.image.incomplete.reload</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/html/canvas/resources/canvas-tests.js"/>
<link href="/html/canvas/resources/canvas-tests.css" rel="stylesheet"/>
</head><body class="show_output">

<h1>2d.pattern.image.incomplete.reload</h1>
<p class="desc"/>


<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"/>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

var img = document.getElementById('yellow.png');
img.src = '../images/red.png';
// This triggers the "update the image data" algorithm,
// and resets the image to the "unavailable" state.
// The image will not go to the "completely available" state
// until a fetch task in the networking task source is processed,
// so the image must not be fully decodable yet:
_assertSame(ctx.createPattern(img, 'repeat'), null, "ctx.createPattern(img, 'repeat')", "null");


});
</script>
<img class="resource" id="yellow.png" src="/images/yellow.png"/>
<img class="resource" id="red.png" src="/images/red.png"/>

</body></html>
